Remove action and container file. refactor into questionlist
