+++ client/src/components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Overview from './Overview';
-import QListContainer from '../containers/QListContainer';
+import QuestionList from './QuestionList';
 import RelatedProducts from './RelatedProducts';
 import Ratings from './Ratings';
 
@@ -15,7 +15,7 @@ function App() {
       <Setter />
       <Overview />
       <RelatedProducts />
-      <QListContainer />
+      <QuestionList />
       <Ratings />
     </>
   );
