+++ client/src/components/QuestionList.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-var */
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { connect, useSelector, useDispatch } from 'react-redux';
 import Question from './Question';
 import QModal from './QModal';
 import { sortQtoAs } from './QAhelperFunctions';
@@ -50,4 +50,20 @@ function QuestionList({ productQs, qFilter, qExpandedBy }) {
   );
 }
 
-export default QuestionList;
+const mapStateToProps = (store) => ({
+  qExpandedBy: store.qList.qExpandedBy,
+  qModal: store.qList.qModal,
+  qFilter: store.qList.qFilter,
+  productQs: store.qList.productQs,
+});//connects the prop to the state saved in the store
+
+//makes it so you can change state depending on inputs   event handler => action/reducer => store
+const mapDispatchToProps = (dispatch) => ({
+  moreAnsweredQuestionsClick: () => dispatch(expandQuestions(true)), //link more question button event
+  onAddAnswerClick: () => dispatch(showQModal(true)), //link add answer event
+});
+
+
+const QuestionListContainer = connect(mapStateToProps, mapDispatchToProps)(QuestionList);
+
+export default QuestionListContainer;
